deps: update rustls, tokio-rustls, quinn, webpki-roots.
This commit updates the Rustls dependency to the recently released
0.21.0 tag, and to update tokio-rustls to 0.24.0 to use the same rustls
release. It also updates webpki-roots to 0.23.0.

It also introduces a temporary patch to update quinn
to a releases using the same 0.21.0 rustls dependency. This updated
dependency isn't ready yet so a patch is used for now.

A couple small upstream API changes are addressed to ensure the build
and tests pass.