svcb: update docs to ref RFC 9460
Since the time this code landed, RFC 9460[0] was published:

  Service Binding and Parameter Specification via the DNS (SVCB and HTTPS
  Resource Records)

This is the definitive reference for SVCB and HTTPS RRs.

It seems as though the implementation did not change meaningfully and so
this commit can largely just update documentation references and copied
quotations to match RFC 9460.

Unfortunately, the Encrypted Client Hello (ECH) aspects of the draft
were removed pre-publication and the `SvcParamKeys` initial IANA
registry left a "reserved" allocation for the `ech` key, but no details
on its use.

Since the code in `svcb.rs` also concerned itself with ECH it now
references draft-ietf-tls-svcb-ech-01[1] where the ECH specific usage of
service parameter is under specification.

There are a handful of small implementation divergences w.r.t ECH that
will be addressed in following commits.

[0]: https://datatracker.ietf.org/doc/html/rfc9460
[1]: draft-ietf-tls-svcb-ech-01