proto: correct ECH service parameter key
Previously `"echconfig"` was being used as the encrypted client hello
(ECH) service parameter key for SVCB/HTTPS RRs.

In RFC960 the parameter key is specified in the intial IANA registry
contents as `"ech"`[0].

This commit updates the two relevant parts of hickory (and corresponding
test data) to use the up-to-date parameter key.

This is a breaking change, however given the very low adoption of ECH,
and the use of the correct `"ech"` key in popular test servers, it
doesn't seem worth trying to maintain backwards compatibility with
earlier draft RFC values.

[0]: https://datatracker.ietf.org/doc/html/rfc9460#section-14.3.2